Find packages rather than define them
This allows `pip wheel .` to succeed. The reason for this is because
the egg_info command is the first setuptools command run by pip. It
runs dist_info on the package, which corrects the missing lib in the
egg_info run. All that is to say this fix doesn't prevent normal
installation AFAICT and does allow for the wheel to be produced.